+++ Blurred-ios/Cells/Comment/CommentCellDelegate.swift
@@ -12,4 +12,5 @@ protocol CommentCellDelegate {
     func likeButtonTapped(commentId: Int, indexPath: IndexPath, reply: Bool)
     func readMoreButtonTapped(commentId: Int, indexPath: IndexPath)
     func moreButtonTapped(commentId: Int, indexPath: IndexPath, reply: Bool)
+    func cellAvatarTapped(commentId: Int, indexPath: IndexPath, reply: Bool, name: String, isReported: Bool, avatarUrl: String)
 }