+++ Blurred-ios/Channels/Video/Overlays/Side Overlay/OverlayViewDelegate.swift
@@ -7,9 +7,10 @@
 //
 
 import Foundation
+import Nuke
 
 protocol ChannelVideoOverlayViewDelegate: AnyObject {
-    func didTapChannel(_ view:ChannelVideoOverlayView, videousername: String)
+    func didTapChannel(_ view:ChannelVideoOverlayView, videousername: String, resizedImageProcessor: [ImageProcessing], isReported: Bool, isBlocked: Bool, name: String)
     func didTapComments(_ view:ChannelVideoOverlayView, videoid: Int)
     func didTapShare(_ view:ChannelVideoOverlayView, videoUrl: String, videoId: Int)
 }